extract shared config for setup command
