+++ cmd/setup/iam.go
@@ -4,54 +4,12 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"os"
 
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/config"
 	"github.com/aws/aws-sdk-go-v2/service/iam"
 	"github.com/aws/aws-sdk-go-v2/service/iam/types"
 )
-
-const (
-	defaultRoleName           = "create-thumbnails-lambda-role"
-	defaultBucketNameOriginal = "original.images.mububoki"
-	defaultBucketNameThumbnail = "thumbnail.images.mububoki"
-	lambdaBasicExecutionRoleARN = "arn:aws:iam::aws:policy/service-role/AWSLambdaBasicExecutionRole"
-	s3PolicyName               = "create-thumbnails-s3-access"
-)
-
-type trustPolicy struct {
-	Version   string            `json:"Version"`
-	Statement []trustStatement  `json:"Statement"`
-}
-
-type trustStatement struct {
-	Effect    string          `json:"Effect"`
-	Principal trustPrincipal  `json:"Principal"`
-	Action    string          `json:"Action"`
-}
-
-type trustPrincipal struct {
-	Service string `json:"Service"`
-}
-
-type s3Policy struct {
-	Version   string        `json:"Version"`
-	Statement []s3Statement `json:"Statement"`
-}
-
-type s3Statement struct {
-	Effect   string   `json:"Effect"`
-	Action   []string `json:"Action"`
-	Resource string   `json:"Resource"`
-}
-
-func envOrDefault(key, defaultValue string) string {
-	if v := os.Getenv(key); v != "" {
-		return v
-	}
-	return defaultValue
-}
 
 func createIAMRole() error {
 	ctx := context.Background()
